Problem date fixed !!!

+++ app/src/main/java/galaxy/pegazus/bienetre/data/LogDay.java
@@ -18,7 +18,7 @@ public class LogDay {
     }
 
     private Integer _id;
-    private Integer date;
+    private Long date;
     private String comment;
     private Integer assessment;
 
@@ -26,11 +26,11 @@ public class LogDay {
         return comment;
     }
 
-    public Integer getDate() {
+    public Long getDate() {
         return date;
     }
 
-    public void setDate(Integer date) {
+    public void setDate(Long date) {
         this.date = date;
     }
